==> Completed steps/5_Concat_inference_files.ipynb <==
import polars as pl

# Match chữ số: inference_0.parquet, inference_1.parquet, ..., inference_9.parquet
df = pl.scan_parquet("train_data/inference_[0-9]*.parquet")

df.sink_parquet("train_data/inference_final.parquet")
